% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_chk.R
\name{all_equal_bycol}
\alias{all_equal_bycol}
\title{Applies the all.equal function to each column between 2 data.tables.}
\usage{
all_equal_bycol(dt.target, dt.current)
}
\arguments{
\item{dt.target}{A target \code{data.table}.}

\item{dt.current}{Another \code{data.table} to be compared with target.}
}
\value{
Either TRUE or a \code{list} describing the differences between
        target and current.
}
\description{
Applies the all.equal function to each column between 2 data.tables.
}
\author{
Yoann Pageaud.
}
