% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_chk.R
\name{allNA.col}
\alias{allNA.col}
\title{Check if any column contains exclusively NAs.}
\usage{
allNA.col(data)
}
\arguments{
\item{data}{An \code{array}, a \code{dataframe} or a \code{matrix}.}
}
\value{
A \code{logical}, TRUE if the array has at least 1 column contaning
        exclusively NAs, FALSE if it has none.
}
\description{
Check if any column contains exclusively NAs.
}
\author{
Yoann Pageaud.
}
