% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_chk.R
\name{best.merged.dt}
\alias{best.merged.dt}
\title{Looks for the best merging operation(s) between two data.tables trying a set
of columns from the second one.}
\usage{
best.merged.dt(dt.x, dt.y, by.x, try.y = NULL, skip.incompatible.type = FALSE)
}
\arguments{
\item{dt.x}{A \code{data.table}.}

\item{dt.y}{Another \code{data.table}.}

\item{by.x}{A \code{character} specifying a single column name from 'dt.x'.}

\item{try.y}{A \code{character} vector specifying multiple
column names from 'dt.y' to be tried for the
merging. If NULL, all columns from 'dt.y' will
be tried for the merging
(Default: try.y = NULL). Columns with a
different type from the one specified in 'by.x'
will raise an error by default
(See 'skip.incompatible.type').}

\item{skip.incompatible.type}{A \code{logical} specifying whether potential
'dt.y' columns of incompatible type specified
in 'try.y' should be automatically skipped
(skip.incompatible.type = TRUE) or not
(skip.incompatible.type = FALSE).}
}
\value{
A \code{list} containing:
         \itemize{
          \item{'best.merged.dt': a \code{data.table} resulting of the best
                merging operation if a single merging operation performed
                the best. If multiple merging operations gave best results,
                operation names are given as a \code{character} vector.}
          \item{'merging.results': a \code{list} of the merging operation
                results. Each result contains 2 elements:
                \itemize{
                 \item{'merge.res': the \code{data.table} resulting from the
                       merging operation.}
                 \item{'NA.count': an \code{integer} vector giving the number
                       of NAs contained in each columns from 'dt.y' after the
                       merging.}
                }
          }
         }
}
\description{
Looks for the best merging operation(s) between two data.tables trying a set
of columns from the second one.
}
\author{
Yoann Pageaud.
}
