% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fun.R
\name{dt.combination}
\alias{dt.combination}
\title{Combines 2 columns from a datatable into a 1 column data.table.}
\usage{
dt.combination(DT, cols, mrg.col, keep.colname = NULL)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{cols}{A \code{character} vector of length 2 matching columns names
of DT.}

\item{mrg.col}{A \code{character} to be used to name the resulting combined
column.}

\item{keep.colname}{An \code{integer} equals to 1, or 2, or NULL. If equals
to 1, the resulting combined column will be named after
the first column in 'cols'. If equals to 2, the resulting
combined column will be named after the second column in
'cols'. If NULL, keep.colname is not used for the naming
of the resulting combined column
(Default: keep.colname = NULL).}
}
\value{
A \code{data.table} with 1 column resulting from the merging of the
        partially duplicated columns of the input.
}
\description{
Combines 2 columns from a datatable into a 1 column data.table.
}
\author{
Yoann Pageaud.
}
\keyword{internal}
