% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fun.R
\name{dt.combine}
\alias{dt.combine}
\title{Combines information using duplicated colnames of a data.table resulting from
merge().}
\usage{
dt.combine(DT)
}
\arguments{
\item{DT}{A \code{data.table} resulting from a merge() operation. Partially
duplicated columns (some values are duplicated at some position of
columns, while at other positions NA values are present in only one
of the columns) are automatically detected using their colnames
suffixes '.x' and '.y', and combined into new columns (thus
reducing the amount of missing values). original duplicated columns
are then removed.}
}
\value{
A \code{data.table} with duplicated columns removed, and resulting
        combined columns appended on the right.
}
\description{
Combines information using duplicated colnames of a data.table resulting from
merge().
}
\references{

}
\author{
Yoann Pageaud.
}
