% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fun.R
\name{dt.int64tochar}
\alias{dt.int64tochar}
\title{Converts columns of 'double.integer64' type into 'character' type in-place.}
\usage{
dt.int64tochar(DT, column.names = NULL)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{column.names}{A \code{character} vector containing column names you
want to convert from 'double.integer64' type to
'character'.
If no column names are specified, the function read
through the entire data.table
(Default: column.names = NULL).}
}
\value{
A \code{data.table}.
}
\description{
Converts columns of 'double.integer64' type into 'character' type in-place.
}
\author{
Yoann Pageaud.
}
