% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fun.R
\name{dt.ls2c}
\alias{dt.ls2c}
\title{Replaces data.table columns of type list to a column of type vector.}
\usage{
dt.ls2c(DT, column.names = NULL)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{column.names}{A \code{character} vector containing column names you
want to convert from list type to vectors.}
}
\value{
A \code{data.table}.
}
\description{
Replaces data.table columns of type list to a column of type vector.
}
\references{

}
\author{
Yoann Pageaud.
}
