% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fun.R
\name{dt.rm.allNA}
\alias{dt.rm.allNA}
\title{Removes columns exclusively containing NAs from a data.table.}
\usage{
dt.rm.allNA(DT, ignore = NULL)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{ignore}{A \code{character} or \code{integer} vector specifying columns
that should be ignored during duplication removal.}
}
\value{
A \code{data.table}.
}
\description{
Removes columns exclusively containing NAs from a data.table.
}
\author{
Yoann Pageaud.
}
