% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fun.R
\name{dt.rm.dup}
\alias{dt.rm.dup}
\title{Removes duplicated column content in a data.table.}
\usage{
dt.rm.dup(DT, ignore = NULL)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{ignore}{A \code{character} or \code{integer} vector specifying columns
that should be ignored during duplication removal.}
}
\value{
A \code{data.table}.
}
\description{
Removes duplicated column content in a data.table.
}
\references{

}
\author{
Yoann Pageaud.
}
