% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_chk.R
\name{allNA.col}
\alias{allNA.col}
\title{Checks if any column contains exclusively NAs and returns colnames if any.}
\usage{
allNA.col(data)
}
\arguments{
\item{data}{An \code{array}, a \code{dataframe} or a \code{matrix}.}
}
\value{
A \code{list} with 'fullNA.col' containing the column names that
        contain exclusively NAs, and 'is.fullNA' a logical TRUE if the array
        has at least 1 column contaning exclusively NAs, FALSE if it has
        none.
}
\description{
Checks if any column contains exclusively NAs and returns colnames if any.
}
\author{
Yoann Pageaud.
}
