% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_fun.R
\name{dt.rm.dup}
\alias{dt.rm.dup}
\title{Removes duplicated column content from a data.table.}
\usage{
dt.rm.dup(DT, ignore = NULL)
}
\arguments{
\item{DT}{A \code{data.table}.}

\item{ignore}{A \code{character} or \code{integer} vector specifying columns
that should be ignored during duplication removal.}
}
\value{
A \code{data.table}.
}
\description{
Removes duplicated column content from a data.table.
}
\examples{
dtbl<-data.table(
  col1 = rep(x = c("hello", "world"), 8),
  col2 = rep(x = c("hello", "world"), 8)) #'col2' is a duplicate of 'col1'.
dt.rm.dup(DT = dtbl) #Only 'col1' remains.
#You can ignore specific columns that will not be remove if duplicated:
dt.rm.dup(DT = dtbl,ignore = "col2")
}
\author{
Yoann Pageaud.
}
